moved db requests to dedicated directory
