+++ api/controllers/sessions/processSearchResults/processResults.js
@@ -2,8 +2,8 @@ const localWebDriver = require('../webDriver');
 const driver = localWebDriver.get_driver();
 const { next_btn_status, test_page } = require('./processResultsActions/nextBtnStatus');
 const { create_session_detail } = require('./processResultsActions/createSessionDetail');
-const { get_user_kw } = require('./processResultsActions/getUserKw');
-const { get_processed_job_ids } = require('./processResultsActions/getProcessedJobIds');
+const { get_user_kw } = require('./processResultsActions/dataBaseRequests/getUserKw');
+const { get_processed_job_ids } = require('./processResultsActions/dataBaseRequests/getProcessedJobIds');
 const { grab_page_tJ_ids } = require('./processResultsActions/grabPageTotalJobIds');
 const { grab_job_url } = require('./processResultsActions/grabJobUrl');
 const { open_job_add } = require('./processResultsActions/openJobAdd');
@@ -44,7 +44,6 @@ exports.process_results = async (userId)  => {
             }
         }
 
-
         let jobData = await grab_all_job_data(userId, jobUrl);
         if (!(jobData.success)) {
             return {
@@ -55,7 +54,7 @@ exports.process_results = async (userId)  => {
         // generate found key words
         let jD = jobData.job_info.job_desc;
         let jDUpperCase = jD.toUpperCase();
-        console.log(jobData.job_info)
+        console.log(jobData.job_info.job_title)
         // check desirability, log and apply accordingly
 
         totalProcessed++
